Refactoring code to allow custom plot sampling
- FFTWidget is now just a PlotWidget (with some added bells and whistles)
- Add a custom FFTSampler class
