+++ src/fft_widget.hpp
@@ -9,7 +9,23 @@
 #include "fft/include/simple_fft/fft_settings.h"
 #include "fft/include/simple_fft/fft.h"
 
+#include "data_series.hpp"
+#include "plot_curve.hpp"
 #include "plot_panner.hpp"
+
+
+class FFTCurveUpdater : public PlotCurveUpdater
+{
+    Q_OBJECT
+
+public:
+    FFTCurveUpdater(DataSeries &data_series);
+
+public slots:
+    virtual void updateCurveSamples(double t_min, double t_max, unsigned int n_pixels) override;
+};
+
+
 
 class FFTWidget : public PlotWidget
 {
@@ -19,12 +35,9 @@ public:
     FFTWidget();
     virtual ~FFTWidget();
 
-signals:
-
-public slots:
-
 protected:
     virtual bool isCurveTrackingEnabled(void) const override { return false; }
+    virtual PlotCurveUpdater* generateNewWorker(QSharedPointer<DataSeries> series) override;
 };
 
 #endif // FFT_WIDGET_HPP
