+++ src/plot_widget.cpp
@@ -1317,7 +1317,10 @@ bool PlotWidget::addSeries(QSharedPointer<DataSeries> series, int axis_id)
     }
 
     // Create a new PlotCurve for the DataSeries
-    PlotCurve *curve = new PlotCurve(series);
+    PlotCurveUpdater* worker = generateNewWorker(series);
+    PlotCurve *curve = new PlotCurve(series, worker);
+
+    qDebug() << "created new curve:" << series << worker;
 
     curve->attach(this);
     curve->setYAxis(axis_id);
@@ -1575,4 +1578,14 @@ void PlotWidget::editAxisScale(QwtPlot::Axis axisId)
 int PlotWidget::getHorizontalPixels() const
 {
     return width();
+}
+
+
+/**
+ * @brief PlotWidget::generateNewWorker - Create a new curve sampling worker
+ * @return
+ */
+PlotCurveUpdater* PlotWidget::generateNewWorker(QSharedPointer<DataSeries> series)
+{
+    return new PlotCurveUpdater(*series);
 }
